# doc/OpenAPI草案.yaml
openapi: 3.0.3
info:
  title: 育英星宠 API
  version: 0.1.0
  description: >
    第一阶段 OpenAPI 草案。基于 React + Node.js + MySQL + WebSocket 架构，
    支持 Admin / Display 双端、固定分值积分规则、15 分钟 Display 解锁、即时领取兑换、
    AI 学情手动生成。
servers:
  - url: /api/v1
...
paths:
...
  /students:
    get:
      tags: [Students]
      summary: 学生列表
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: classId
          schema:
            type: integer
            format: int64
        - in: query
          name: keyword
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
  /students/{id}:
    get:
      tags: [Students]
      summary: 学生详情
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
  /students/import:
    post:
      tags: [Students]
      summary: 批量导入学生
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                classId:
                  type: integer
                  format: int64
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
  schemas:
    BaseResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
        data:
          nullable: true
...